It has been changed

+++ src/main/java/ru/kata/spring/boot_security/demo/services/RoleService.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface RoleService {
     List<Role> getSetOfRoles(List<String> rolesId);
-    Role getRoleById(long id);
+    Role getRoleById(int id);
     List<Role> getRoleList();
 }
